fix(lerna): fix grapher tests not running
