+++ packages/@ourworldindata/grapher/tsconfig.test.json
@@ -1,5 +1,5 @@
 {
-    "include": ["src/**/*.test.tsx", "src/**/*.d.ts"],
+    "include": ["src/**/*", "stories.jsdom.test"],
     "compilerOptions": {
         "target": "es2016" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
         "jsx": "react" /* Specify what JSX code is generated. */,
